backend/detection_service/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
ultralytics
opencv-python
numpy
torch
torchvision
crewai

